Rename module to "subscriptions-core"
