+++ app/build.gradle
@@ -318,7 +318,7 @@ dependencies {
     implementation project(':components:icons')
     implementation project(':components:chartview')
 
-    implementation project(':subscriptions-kit')
+    implementation project(':subscriptions-core')
     implementation project(':subscriptions-google-play')
 
     // Integration with activities
