+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/usersubscription/BuySubscriptionViewModel.kt
@@ -3,8 +3,8 @@ package io.horizontalsystems.bankwallet.modules.usersubscription
 import android.app.Activity
 import androidx.lifecycle.viewModelScope
 import io.horizontalsystems.bankwallet.core.ViewModelUiState
-import io.horizontalsystems.subscriptionskit.SubscriptionPlan
-import io.horizontalsystems.subscriptionskit.UserSubscriptionManager
+import io.horizontalsystems.subscriptions.core.SubscriptionPlan
+import io.horizontalsystems.subscriptions.core.UserSubscriptionManager
 import kotlinx.coroutines.launch
 
 class BuySubscriptionViewModel : ViewModelUiState<BuySubscriptionUiState>() {
