+++ settings.gradle
@@ -3,5 +3,5 @@ include ':app'
 include ':core'
 include ':components:icons'
 include ':components:chartview'
-include ':subscriptions-kit'
+include ':subscriptions-core'
 include ':subscriptions-google-play'
