+++ subscriptions-core/.gitignore
