+++ subscriptions-core/proguard-rules.pro
