+++ subscriptions-core/src/main/AndroidManifest.xml
