+++ subscriptions-core/src/main/java/io/horizontalsystems/subscriptions/core/SubscriptionPlan.kt
@@ -1,3 +1,3 @@
-package io.horizontalsystems.subscriptionskit
+package io.horizontalsystems.subscriptions.core
 
 data class SubscriptionPlan(val id: String, val name: String)
