+++ subscriptions-core/src/test/java/io/horizontalsystems/subscriptions/core/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.subscriptionskit
+package io.horizontalsystems.subscriptions.core
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
